Combine pdo.php and pdocore.php
These were originally split up because the functions in pdo.php wanted to log
errors to the database, which could not happen until pdocore.php established a
database connection.  Now that we no longer log errors to the database, this
distinction is no longer necessary.
